cairosink: rework autotools use
We now check for the existence of cairo-glx or cairo-egl + cairo-glesv2 through
pkg-config instead of checking for glx/egl directly, which was wrong because
having one of them does not guarantee the presence of the corresponding cairo
backend.
Also introduced the defines USE_CAIRO_EGL and USE_CAIRO_GLX to know what
backend(s) to build.