threading: replace last_data system with simpler boolean
This change simplifies code that was uselessly complicated and prevent a race
condition when the same last_data (e.g. from a GstThreadInfo) was used by two
callers from different threads at about the same time. This has been stopped by
removing the last_data member from GstThreadInfo, a gboolean pointer on the
stack is used instead (and for that gst_thread_invoke_sync() was changed from a
macro to a function).
This changed the definition of GstThreadInfo and the prototype of
gst_threading_main_context_invoke_sync(), with last_data removed.